add cosine scheduler option
